% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_estimation.R
\name{compute_L}
\alias{compute_L}
\title{Compute Length at Age L(t)}
\usage{
compute_L(age, Linf, L0, k, growth_model = c("vb", "gompertz", "logistic"))
}
\arguments{
\item{age}{Numeric vector. Ages.}

\item{Linf}{Numeric. Asymptotic length.}

\item{L0}{Numeric. Length at birth.}

\item{k}{Numeric. Native growth coefficient.}

\item{growth_model}{Character. Growth model.}
}
\value{
Numeric vector of predicted lengths.
}
\description{
Compute Length at Age L(t)
}
