% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_estimation.R
\name{.enforce_constraints}
\alias{.enforce_constraints}
\title{Enforce Biological Constraints on Parameter Draws}
\usage{
.enforce_constraints(par_draws)
}
\arguments{
\item{par_draws}{data.table with Linf, L0, Lmat, tmat columns.}
}
\value{
data.table with constraints enforced.
}
\description{
Enforce Biological Constraints on Parameter Draws
}
\keyword{internal}
