% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_estimation.R
\name{is_stanfit}
\alias{is_stanfit}
\title{Check if Object is a CmdStanMCMC Fit}
\usage{
is_stanfit(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical.
}
\description{
Check if Object is a CmdStanMCMC Fit
}
\keyword{internal}
