% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_functions.R
\name{print.vitalBayes_ppc}
\alias{print.vitalBayes_ppc}
\title{Print PPC Summary}
\usage{
\method{print}{vitalBayes_ppc}(x, ...)
}
\arguments{
\item{x}{A vitalBayes_ppc object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print PPC Summary
}
