% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_estimation.R
\name{compute_tmax}
\alias{compute_tmax}
\title{Compute Maximum Age from Growth Parameters}
\usage{
compute_tmax(
  Linf,
  L0,
  k,
  growth_model = c("vb", "gompertz", "logistic"),
  Linf_factor = 0.99
)
}
\arguments{
\item{Linf}{Numeric. Asymptotic length.}

\item{L0}{Numeric. Length at birth.}

\item{k}{Numeric. Native growth coefficient.}

\item{growth_model}{Character. Growth model.}

\item{Linf_factor}{Numeric in (0,1). Fraction of Linf (default 0.99).}
}
\value{
Numeric tmax value.
}
\description{
Estimates tmax as age when L(t) reaches a specified fraction of Linf.
}
\examples{
tmax <- compute_tmax(Linf = 126, L0 = 35, k = 0.016,
                     growth_model = "vb", Linf_factor = 0.99)

}
