% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_estimation.R
\name{M_lorenzen}
\alias{M_lorenzen}
\title{Lorenzen Natural Mortality Model}
\usage{
M_lorenzen(
  age,
  Linf,
  L0,
  Lmat,
  tmat,
  lw_fun = NULL,
  weight_based = FALSE,
  growth_model = c("vb", "gompertz", "logistic"),
  sample_params = TRUE
)
}
\arguments{
\item{age}{Numeric vector of ages.}

\item{Linf}{Asymptotic length.}

\item{L0}{Length at birth.}

\item{Lmat}{Length at maturity.}

\item{tmat}{Age at maturity.}

\item{lw_fun}{Length-weight function (required if weight_based = TRUE).}

\item{weight_based}{Use weight-based formulation?}

\item{growth_model}{Growth model for L(t).}

\item{sample_params}{Sample parameters from their distributions?}
}
\value{
Numeric vector of instantaneous mortality rates.
}
\description{
Lorenzen Natural Mortality Model
}
