% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_estimation.R
\name{M_peterson_wroblewski}
\alias{M_peterson_wroblewski}
\title{Peterson-Wroblewski Natural Mortality Model}
\usage{
M_peterson_wroblewski(
  age,
  Linf,
  L0,
  Lmat,
  tmat,
  lw_fun,
  growth_model = c("vb", "gompertz", "logistic")
)
}
\arguments{
\item{age}{Numeric vector of ages.}

\item{Linf}{Asymptotic length.}

\item{L0}{Length at birth.}

\item{Lmat}{Length at maturity.}

\item{tmat}{Age at maturity.}

\item{lw_fun}{Length-weight function.}

\item{growth_model}{Character. Growth model for L(t).}
}
\value{
Numeric vector of instantaneous mortality rates.
}
\description{
Peterson-Wroblewski Natural Mortality Model
}
